+++ src/mysql/CsvBuffer.cc
@@ -370,236 +370,4 @@ shared_ptr<CsvBuffer> newCsvStreamBuffer(shared_ptr<CsvStream> const& csvStream)
     return make_shared<CsvStreamBuffer>(csvStream);
 }
 
-TransferTracker::Ptr TransferTracker::_globalMt;
-
-void TransferTracker::setup(string const& transferMethodStr, std::size_t max, string const& directory,
-                            std::size_t minMBInMem, std::size_t maxResultTableSizeBytes) {
-    if (_globalMt != nullptr) {
-        throw util::Bug(ERR_LOC, "MemoryTracker::setup called when MemoryTracker already setup!");
-    }
-    TransferMethod tm = transferMethodFromString(transferMethodStr);
-    if (tm == MEMDISK) {
-        if (!verifyDir(directory)) {
-            throw util::Bug(ERR_LOC, "MemoryTracker::setup called with bad directory! " + directory);
-        }
-    }
-    size_t const MB_SIZE_BYTES = 1024 * 1024;
-    if (tm == MEMORY && max < maxResultTableSizeBytes) {
-        throw util::Bug(ERR_LOC,
-                        "Configuration error resultdb maxTransferMemMB=" + to_string(max / MB_SIZE_BYTES) +
-                                " must be larger than maxtablesize_mb= " +
-                                to_string(maxResultTableSizeBytes / MB_SIZE_BYTES));
-    }
-    _globalMt = TransferTracker::Ptr(new TransferTracker(tm, max, directory, minMBInMem));
-}
-
-bool TransferTracker::verifyDir(string const& dirName) {
-    sfs::path dir = dirName;
-    if (!(sfs::exists(dir) && sfs::is_directory(dir))) {
-        LOGS(_log, LOG_LVL_ERROR, "verifyDir, " + dirName + " is not a valid directory");
-        return false;
-    }
-    return true;
-}
-
-TransferTracker::TransferMethod TransferTracker::transferMethodFromString(string const& strType) {
-    string str;
-    for (unsigned char c : strType) {
-        str += tolower(c);
-    }
-    TransferMethod tMethod;
-    if (str == "memory") {
-        tMethod = MEMORY;
-        LOGS(_log, LOG_LVL_INFO, "Result TransferMethod set to memory");
-    } else if (str == "stream") {
-        tMethod = STREAM;
-        LOGS(_log, LOG_LVL_INFO, "Result TransferMethod set to stream");
-    } else if (str == "memdisk") {
-        tMethod = MEMDISK;
-        LOGS(_log, LOG_LVL_INFO, "Result TransferMethod set to memdisk");
-    } else {
-        tMethod = MEMORY;
-        LOGS(_log, LOG_LVL_ERROR,
-             "Result TransferMethod set to memory due to invalid string '"
-                     << strType << "'"
-                     << " valid entries are 'memory', 'stream', 'memdisk'");
-    }
-    return tMethod;
-}
-
-TransferTracker::MemoryRaii::Ptr TransferTracker::createRaii(size_t fileSize, bool wait) {
-    unique_lock ulck(_mtx);
-    if (wait) {
-        if (fileSize > _max) {
-            throw util::Bug(ERR_LOC, "MemoryTracker::createRaii file too large " + to_string(fileSize) +
-                                             " max=" + to_string(_max));
-        }
-        _cv.wait(ulck, [this, fileSize]() { return (fileSize + _total) < _max; });
-    }
-    MemoryRaii::Ptr pRaii(new MemoryRaii(fileSize));
-    return pRaii;
-}
-
-void TransferTracker::_incrTotal(size_t sz) {
-    // _mtx must already be locked.
-    _total += sz;
-    _cv.notify_one();  // Many items may be waiting on a large file, so there may be room for more
-}
-
-void TransferTracker::_decrTotal(size_t sz) {
-    lock_guard ulck(_mtx);
-    if (sz > _total) {
-        throw util::Bug(ERR_LOC,
-                        "MemoryTracker::_decrTotal sz=" + to_string(sz) + " > total=" + to_string(_total));
-    }
-    _total -= sz;
-    _cv.notify_one();
-}
-
-void CsvStrMem::waitReadyToRead() {
-    auto memTrack = TransferTracker::get();
-    if (memTrack == nullptr) {
-        throw util::Bug(ERR_LOC, "CsvStrMem::waitReadyToRead MemoryTracker is NULL");
-    }
-    bool const wait = true;
-    _memRaii = memTrack->createRaii(_expectedBytes, wait);
-
-    // Read directly without starting a separate thread.
-    _csvLambda();
-}
-
-void CsvStrMem::push(char const* data, size_t size) {
-    if (_cancelled) return;
-    _bytesRead += size;
-    if (data != nullptr && size != 0) {
-        _records.emplace_back(make_shared<string>(data, size));
-    } else {
-        // Empty string is meant to indicate the end of the stream.
-        _records.emplace_back(make_shared<string>());
-    }
-}
-
-shared_ptr<string> CsvStrMem::pop() {
-    shared_ptr<string> front = _records.front();
-    _records.pop_front();
-    return front;
-}
-
-CsvStrMemDisk::CsvStrMemDisk(std::size_t expectedBytes, QueryId qId, UberJobId ujId)
-        : CsvStrMem(expectedBytes), _qId(qId), _ujId(ujId) {
-    auto memTrack = TransferTracker::get();
-    if (memTrack == nullptr) {
-        throw util::Bug(ERR_LOC, "CsvStrMemDisk constructor MemoryTracker is NULL");
-    }
-    sfs::path fPath = memTrack->getDirectory();
-    string fileName = memTrack->getBaseFileName() + "_" + to_string(_qId) + "_" + to_string(ujId);
-    fPath /= fileName;
-    _filePath = fPath;
-
-    _minBytesInMem = memTrack->getMinBytesInMem();
-}
-
-void CsvStrMemDisk::waitReadyToRead() {
-    auto memTrack = TransferTracker::get();
-    if (memTrack == nullptr) {
-        throw util::Bug(ERR_LOC, "CsvStrMemDisk::waitReadyToRead MemoryTracker is NULL");
-    }
-    bool const nowait = false;
-    _memRaii = memTrack->createRaii(_expectedBytes, nowait);
-
-    // Read directly without starting a separate thread.
-    _csvLambda();
-}
-
-bool CsvStrMemDisk::_mustWriteToTmpFile() {
-    // Once writing to file, this instance must keep writing to file.
-    if (_writingToFile) return true;
-
-    auto memTrack = TransferTracker::get();
-    // If too much memory is being used for transfers, star writing large transfers to files.
-    if (memTrack->getTotal() > memTrack->getMax()) {
-        if (_records.size() > _minRecordsSize && _bytesRead > _minBytesInMem) {
-            _writingToFile = true;
-        }
-    }
-    return _writingToFile;
-}
-
-void CsvStrMemDisk::push(char const* data, size_t size) {
-    // Push is always ok, no need to wait.
-    if (_cancelled) return;
-    _bytesRead += size;
-    if (_mustWriteToTmpFile()) {
-        _writeToTmpfile(data, size);
-        return;
-    }
-    if (data != nullptr && size != 0) {
-        _records.emplace_back(make_shared<string>(data, size));
-    } else {
-        // Empty string is meant to indicate the end of the stream.
-        _records.emplace_back(make_shared<string>());
-    }
-}
-
-shared_ptr<string> CsvStrMemDisk::pop() {
-    if (_records.size() > 0) {
-        shared_ptr<string> front = _records.front();
-        _records.pop_front();
-        return front;
-    }
-    return _readFromTmpFile();
-}
-
-void CsvStrMemDisk::_writeToTmpfile(char const* data, std::size_t size) {
-    // Open the file if needed
-    auto oldState = _fState.exchange(OPEN_W);
-    if (oldState == INIT) {
-        _file.open(_filePath, fstream::out);
-    }
-    if (!_file.is_open() || _fState != OPEN_W) {
-        LOGS(_log, LOG_LVL_ERROR,
-             "CsvStrMemDisk::_writeTofile file isn't open " << _filePath << " or bad state=" << _fState);
-        _fileError = true;
-        return;
-    }
-
-    _file.write(data, size);
-    _bytesWrittenToTmp += size;
-}
-
-std::shared_ptr<std::string> CsvStrMemDisk::_readFromTmpFile() {
-    if (_fState == OPEN_W) {
-        _fState = CLOSE_W;
-        _file.close();
-    }
-    auto oldState = _fState.exchange(OPEN_R);
-    if (oldState == CLOSE_W) {
-        _file.open(_filePath, fstream::in);
-        _bytesLeft = _bytesWrittenToTmp;
-    }
-    if (!_file.is_open() || _fState != OPEN_R) {
-        // This is extremely unlikely
-        if (!getContaminated())
-            LOGS(_log, LOG_LVL_ERROR,
-                 "CsvStrMemDisk::_readFromfile file isn't open " << _filePath << " or bad state=" << _fState);
-        setContaminated();
-        return make_shared<string>("$");
-    }
-
-    std::size_t buffSz = std::min(100'000ul, _bytesLeft);
-    auto strPtr = make_shared<string>();
-    strPtr->resize(buffSz);
-    _file.read(strPtr->data(), buffSz);
-    _bytesLeft -= buffSz;
-    return strPtr;
-}
-
-CsvStrMemDisk::~CsvStrMemDisk() {
-    if (_fState != INIT) {
-        LOGS(_log, LOG_LVL_INFO, "~CsvStrMemDisk() remove " << _filePath);
-        _file.close();
-        std::remove(_filePath.c_str());
-    }
-}
-
 }  // namespace lsst::qserv::mysql
